Flytter normaliseringsfunksjoner til egen felles fil
MELOSYS-4390
