+++ src/felleskomponenter/skjema/input/input.js
@@ -56,25 +56,6 @@ InnerInputComponent.defaultProps = {
   onChange: undefined,
 };
 
-export const normalizeInt = (value, previousValue) => {
-  if (value === "") return null;
-
-  const isInt = value.match(/^\d+$/g) !== null;
-  return isInt ? value : previousValue;
-};
-export const normalizeDecimal = (value, previousValue) => {
-  if (value === "") return null;
-
-  const valuePreferDot = value.replace(",", ".");
-  const isIntOrDecimal = valuePreferDot.match(/^\d+([.]\d*)?$/g) !== null;
-
-  return isIntOrDecimal ? valuePreferDot : previousValue;
-};
-
-export const begrensAntallTegn = (antallTegn) => (value) => {
-  return antallTegn && value.length > antallTegn ? value.substr(0, antallTegn) : value;
-};
-
 function Input({ feltNavn, bredde = "fullbredde", normalize = undefined, ...rest }) {
   return (
     <Field bredde={bredde} name={feltNavn} normalize={normalize} component={InnerInputComponent} props={{ ...rest }} />
@@ -84,13 +65,11 @@ function Input({ feltNavn, bredde = "fullbredde", normalize = undefined, ...rest
 Input.propTypes = {
   bredde: PT.string,
   feltNavn: PT.string.isRequired,
-  feltType: PT.oneOf(["desimal", "heltall"]),
   normalize: PT.func,
 };
 
 Input.defaultProps = {
   bredde: "fullbredde",
-  feltType: undefined,
   normalize: undefined,
 };
