+++ src/felleskomponenter/skjema/input/input.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Input, { begrensAntallTegn, InnerInputComponent, normalizeDecimal, normalizeInt } from "./input";
+import Input, { InnerInputComponent } from "./input";
 
 describe("Input", () => {
   let props = null;
@@ -36,58 +36,6 @@ describe("Input", () => {
     const input = shallow(<Input {...props} />);
 
     expect(input.props().normalize).toBe(props.normalize);
-  });
-});
-
-describe("normalizeDecimal", () => {
-  it("returnerer null for tom string", () => {
-    expect(normalizeDecimal("", "5")).toBe(null);
-  });
-
-  it("erstatter komma med punktum", () => {
-    expect(normalizeDecimal("5,", "5")).toBe("5.");
-  });
-
-  it("returnerer value hvis value er desimal", () => {
-    expect(normalizeDecimal("5.", "5")).toBe("5.");
-  });
-
-  it("returnerer value hvis value er heltall", () => {
-    expect(normalizeDecimal("5", null)).toBe("5");
-  });
-
-  it("returnerer previousvalue hvis value ikke er heltall eller desimal", () => {
-    expect(normalizeDecimal("5..", null)).toBe(null);
-    expect(normalizeDecimal(".", null)).toBe(null);
-    expect(normalizeDecimal(".5", null)).toBe(null);
-  });
-});
-
-describe("normalizeInt", () => {
-  it("returnerer null for tom string", () => {
-    expect(normalizeInt("", "5")).toBe(null);
-  });
-
-  it("returnerer previousValue hvis value er desimal", () => {
-    expect(normalizeInt("5.", "5")).toBe("5");
-  });
-
-  it("returnerer value hvis value er heltall", () => {
-    expect(normalizeInt("5", null)).toBe("5");
-  });
-});
-
-describe("begrensAntallTegn", () => {
-  it("returnerer begrenset antall tegn", () => {
-    expect(begrensAntallTegn(5)("12345678911")).toBe("12345");
-  });
-
-  it("begrenser ikke dersom antallet tegn ikke er nådd maksimum", () => {
-    expect(begrensAntallTegn(10)("Heihei")).toBe("Heihei");
-  });
-
-  it("begrenser ikke dersom antallet tegn er like mange som maksimalt tillatte tegn", () => {
-    expect(begrensAntallTegn(18)("her har vi 18 tegn")).toBe("her har vi 18 tegn");
   });
 });
