+++ src/utils/normalisering/index.js
@@ -1,0 +1,19 @@
+export const normalizeInt = (value, previousValue) => {
+  if (value === "") return null;
+
+  const isInt = value.match(/^\d+$/g) !== null;
+  return isInt ? value : previousValue;
+};
+
+export const normalizeDecimal = (value, previousValue) => {
+  if (value === "") return null;
+
+  const valuePreferDot = value.replace(",", ".");
+  const isIntOrDecimal = valuePreferDot.match(/^\d+([.]\d*)?$/g) !== null;
+
+  return isIntOrDecimal ? valuePreferDot : previousValue;
+};
+
+export const begrensAntallTegn = (antallTegn) => (value) => {
+  return antallTegn && value.length > antallTegn ? value.substr(0, antallTegn) : value;
+};
